fix nav scroll on style page
